Created unit test for document serialization and deserialization

+++ src/main/java/dbf0/document/types/DDecimal.java
@@ -21,7 +21,7 @@ public final class DDecimal extends DElement {
   }
 
   @Override public @NotNull DElementType getType() {
-    return DElementType.INT;
+    return DElementType.DECIMAL;
   }
 
   @Override protected int compareToSameType(DElement o) {
